+++ src/main/java/com/myweb/MusicLD/repository/UserRepository.java
@@ -7,11 +7,12 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity,Long> {
+public interface UserRepository extends JpaRepository<UserEntity, BigInteger> {
     @Query("SELECT u from UserEntity u WHERE u.username = ?1")
     Optional<UserEntity> findByUsername(String username);
